fix: minio_test now initialize the configuration to detect the storage class
