+++ nuvolaris/templates/00-minio-pvc.yaml
@@ -26,7 +26,7 @@ metadata:
 spec:
   accessModes:
     - ReadWriteOnce
-  storageClassName: standard
+  storageClassName: {{storage_class}}
   resources:
     requests:
       storage: {{minio_volume_size}}Gi
