+++ tests/kind/mongodb_test.ipy
@@ -34,3 +34,6 @@ assert(kube.get("mdbc/nuvolaris-mongodb"))
 !kubectl get secret -n nuvolaris -ojson nuvolaris-mongodb-nuvolaris-nuvolaris | jq -r '.data | with_entries(.value |= @base64d)'
 
 assert(mdb.delete())
+
+!kubectl -n nuvolaris delete all --all
+!kubectl -n nuvolaris delete pvc --all
